+++ CoreTestament/src/core/Internal/InternalTest.cpp
@@ -10,7 +10,11 @@ InternalTest::InternalTest(std::string name_, FunctionVariant function_)
     : InternalTest(std::move(name_), TestOptions{}, std::move(function_)) {}
 
 InternalTest::InternalTest(std::string name_, TestOptions options_, FunctionVariant function_)
-    : name(std::move(name_)), options(std::move(options_)), function(std::move(function_)) {}
+    : name(std::move(name_)), options(std::move(options_)), function(std::move(function_)) {
+    if (options.isDisabled()) {
+        status = TestStatus::Status::Skipped;
+    }
+}
 
 InternalTest::~InternalTest() = default;